+++ CHANGELOG.rst
@@ -38,6 +38,9 @@ v33.2.0 (unreleased)
   unit paths when available from elfs.
   https://github.com/nexB/purldb/issues/260
 
+- Extract all archives recursively in the `scan_single_package` pipeline.
+  https://github.com/nexB/scancode.io/issues/1081
+
 v33.1.0 (2024-02-02)
 --------------------